Move ColorValueEditor/LengthValueEditor/ScaleValueEditor/ParamEditor to shared
Moves the remaining "ready now" and "small refactor" editors into
ag-website-shared/src/components/theme-builder:

- ColorValueEditor, LengthValueEditor, ScaleValueEditor - moved as-is,
  no grid-specific content (the ColorValueEditor preventTransparency
  check uses real theming-engine param names, not app-specific
  curation, so it was already portable)
- FontFamilyValueEditor, ImageValueEditor - the curated font list and
  docs URL are genuine per-host editorial choices, so these now expose
  a configure-once hook (setFontFamilyOptions/setImageValuesDocsUrl),
  mirroring ParamModel's setParamDocsProvider pattern. Grid registers
  its exact original font list and docs URL in a new
  registerEditorConfig.ts, imported once from ThemeBuilder.tsx.
- ParamEditor - the dispatcher had no remaining grid-specific logic of
  its own; it only stayed local because its dispatch table pointed at
  the editors above. Moves now that all of them are shared.

Also splits themeImport.ts: loadFile/useFileDropZone/FileLoadResult
had zero grid coupling and move to a new shared fileImport.ts;
validateThemeCode/applyValidatedTheme/extractParts stay local since
they're tied to grid's Preset type and PartModel.

Verified: tsc identical to baseline (the only diff is 3 known
pre-existing errors relocating to ParamEditor.tsx's new path), 40/40
tests pass, eslint/prettier clean, and manually in the running dev
server - font family dropdown renders grid's full registered font
list with Google Fonts loading correctly, accent color changes
propagate to the live grid previews, and the "All Parameters" search
correctly dispatches shadow-type params to CssValueEditor.
