+++ documentation/ag-grid-docs/src/components/theme-builder/components/general/GridPreview.tsx
@@ -1,5 +1,6 @@
 import { ColorPicker } from '@ag-website-shared/components/theme-builder/ColorPicker';
 import { withErrorBoundary } from '@ag-website-shared/components/theme-builder/ErrorBoundary';
+import { LoadFontFamilyMenuFonts } from '@ag-website-shared/components/theme-builder/FontFamilyValueEditor';
 import { InfoTooltip } from '@ag-website-shared/components/theme-builder/Tooltip';
 import { useApplicationConfigAtom } from '@ag-website-shared/theming/application-config';
 import { useRenderedTheme } from '@ag-website-shared/theming/rendered-theme';
@@ -28,7 +29,6 @@ import {
 } from 'ag-grid-enterprise';
 import { AgGridReact } from 'ag-grid-react';
 
-import { LoadFontFamilyMenuFonts } from '../editors/FontFamilyValueEditor';
 import { GridConfigDropdownButton } from '../grid-config/GridConfigDropdown';
 import { useGridOptions } from '../grid-config/grid-config-atom';
 import { allPresets } from '../presets/presets';
