+++ external/ag-website-shared/src/components/theme-builder/ScaleValueEditor.tsx
@@ -1,8 +1,8 @@
-import { FormattedInput } from '@ag-website-shared/components/theme-builder/FormattedInput';
-import type { ValueEditorProps } from '@ag-website-shared/components/theme-builder/ValueEditorProps';
-import { stripFloatingPointErrors } from '@ag-website-shared/theming/utils';
-
 import type { ScaleValue } from 'ag-grid-community';
+
+import { stripFloatingPointErrors } from '../../theming/utils';
+import { FormattedInput } from './FormattedInput';
+import type { ValueEditorProps } from './ValueEditorProps';
 
 export const ScaleValueEditor = ({
     value,
